backend/django/djangodb/requirements.txt

Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
requests==2.32.3
mysqlclient==2.1.1
pillow==10.4.0
python-dotenv==1.0.1
# mariadb==1.0.11